Lua: Refactored the Slot object Batches slot change requests so that it is not one per slot, but one per set of changes
-- auto_layout_gen4_gen5.lua
Found Bill's PC memory address is HG/SS in preparation for SL
Refactored main function so that it is easy to access the pokemon data of those stored in boxes
Removed some debugging code
Put in a stop-gap check on level which sometimes is > 100 and throws validation errors in the server
Added code so that it is possible for future streamers to find the start of Bill's PC in memory

-- send_data_to_server.lua
Refactored send function to handle more than one at a time.  This is necessary for when we blit an entire game's worth of pokemon, not just the 6 in the party
Created a basic, single-level json converter for sending requests
Changed content-type of requests from x-www-form-urlencoded to json

Node
-- Config changes
Changed SoulLinkGap style to manual imageWrapperShifts style
Fixed a bunch of issues with getting the config to update with the webpack-dev-server
Fixed a bug from last checkin where the config wasn't actually updating when a config file was touched

-- Server
Fixed some typo bugs
Decreased the dead connection check interval from 5000ms to 2000ms, though why it only detects disconnects sometimes is beyond me
Added infrastructure for handling boxes
Removed unused export from constants
Completely refactored Pokemon Images so that they automatically update when being JSON serialized... this only really matters for the development server, but still
Did a bunch of work on the slot.js of the server... some bugs, some prep for SL
Fixed a bug in validate-argument.js that was causing it not to return the argument value
Fixed odd bug where it was trying to destroy an already destroyed Discord.js client

+++ node/server/soul-link/index.js
@@ -46,9 +46,11 @@ class SoulLink extends EventEmitter {
         c.on('reconnecting', () => console.warn('Bot lost connection to Discord.  Reconnectiong...'));
         c.on('disconnect', () => {
             console.warn('Bot disconnected from Discord.');
-            this._client.destroy();
-            this._client = null;
-            this._initialized = false;
+            if (this._client) {
+                this._client.destroy();
+                this._client = null;
+                this._initialized = false;
+            }
         });
         c.on('error', (err) => {
             console.error(`ERROR in connection to Discord: ${err.name}: ${err.message}`);
